test: Photo4CutServiceTest에서 S3Controller를 쓰는 메소드를 Photo4cutControllerTest로 수정하여 옮김
